renamed component names, added font for splash screen, organized assets
