+++ ui/src/app/App.js
@@ -2,7 +2,7 @@ import React from 'react'
 import { BrowserRouter as Router, Route, Switch } from 'react-router-dom'
 import './App.css'
 import Login from '../login/Login'
-import WishSummary from '../WishSummary'
+import WishList from '../wishList'
 import WatchAuth from '../auth/WatchAuth'
 import ChildInfo from '../childinfo/ChildInfo'
 import CreateWish from '../landing/CreateWish'
@@ -14,7 +14,7 @@ function App() {
         <Switch>
           <Route exact path="/child-info" component={ChildInfo} />
           <Route exact path="/landing" component={CreateWish} />
-          <Route exact path="/wish-summary" component={WishSummary} />
+          <Route exact path="/wish-summary" component={WishList} />
           <Route exact path="/" component={Login} />
         </Switch>
       </WatchAuth>
