+++ ui/src/assets/images/Logo_Galaxy_Colour.png
