+++ ui/src/landing/splashScreen/styles.scss
@@ -6,7 +6,7 @@
   width: 100%;
   background-color: $light-blue;
   color: $dark-blue;
-  font-family: 'Lato', sans-serif;
+  font-family: 'Arial Rounded MT Bold', sans-serif;
   padding: 2.5em;
   box-sizing: border-box;
 
@@ -23,6 +23,13 @@
     max-height: 3em;
   }
 
+  @font-face {
+    font-family: 'Arial Rounded MT Bold';
+    font-style: normal;
+    font-weight: normal;
+    src: local('Arial Rounded MT Bold'), url('../../assets/fonts/ARLRDBD_0.woff') format('woff');
+}
+  
   img.header-galaxy-logo {
     max-height: 10em;
   }
