+++ ui/src/landing/wishTypeCard/index.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import './styles.scss'
 
-export const WishCard = ({imgSrc, altText, title, subtitle}) => (
+export const WishTypeCard = ({imgSrc, altText, title, subtitle}) => (
   <div id="WishCard" className="containerVertical">
     <img src={imgSrc} alt={altText} className="wish-card-image"/>
     <span className="card-title">{title}</span>
