+++ ui/src/landing/wishTypeCard/styles.scss
