+++ ui/src/landing/wishTypeList/index.test.js
