+++ ui/src/landing/wishTypeList/styles.scss
