+++ ui/src/wishList/imgPlaceholder/styles.scss
