+++ ui/src/wishList/index.test.js
