+++ ui/src/wishList/styles.scss
@@ -1,6 +1,6 @@
 @import '../styles/colors.scss';
 
-#WishSummary {
+#wishList {
     padding: 1em;
     background-color: $lighter-gray;
     position: absolute;
