+++ ui/src/wishList/wish/index.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react'
 import './styles.scss'
-import kidImg from '../../assets/icn_To_Be_Astronaut_130x130.png'
-import sponsorImg from '../../assets/icn_To_Be_Astronaut_130x130.png'
+import kidImg from '../../assets/images/icn_To_Be_Astronaut_130x130.png'
+import sponsorImg from '../../assets/images/icn_To_Be_Astronaut_130x130.png'
 import ImgPlaceholder from '../imgPlaceholder'
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 import { faChevronRight } from '@fortawesome/free-solid-svg-icons'
