+++ ui/src/wishList/wish/index.test.js
