+++ ui/src/wishList/wish/styles.scss
