+++ ui/src/wishList/wishFilter/styles.scss
